[FIX] grgit-release now infers version on toString
In 0.8.0, the version wasn't inferred until the task
graph was ready. Since plugins, such as maven-publish,
would try to read the version earlier this resulted in
IllegalStateExceptions.

Without the task graph, the scope and stage are now
provided via project properties which can be accessed
at any point.

The version is now inferred as soon as anything calls
toString.

Correspondingly, the ready* and release* task rules were
replaced with static prepare and release tasks.