Fix #77 Correctly handling tag message in opinion plugin w/ no prev tag.
When running the release-opinion plugin for the first release, the defaulted
tag strategy fails to generate the log used for the tag message. Now it checks
to see if the previous version's tag exists. If it does use the previous
behavior of generating a log between the previous version and the current
HEAD. If it doesn't exist, generate a log of everything up to HEAD.