Reimplementing NearestVersionLocator for parallel.
In order to support #115, sort versions by ancestry
instead of distance, with version precedence being
the tie breaker. A flag is now available on
the locate() method to include tags from parallel
histories. These tags will always be treated as
normal versions, indicating that they are
developing towards that version and that the current
branch should not use it.