added testcases for issue #121
By diving deeper into the current implementation and the testcases,
I found that the currently the *_MAJOR_MINOR_X strategies do only increment
either the major or the minor, but never the patch.

I think if the scope *_MAJOR_MINOR_X is used, every new version should
increment the patch.

This is a breaking change.