fixes #121: *_MAJOR_MINOR_X strategy does always increment the patch
When the *_MAJOR_MINOR_X strategy matches, the patch is always incremented.
This allows to have a branch release/1.2.x to create 1.2.0, 1.2.1, 1.2.2,
... versions.

Without this change, it is possible to do the above with the gradle-git
plugin, but not with the nebula-release plugin, which uses gradle-git under
the covers, but uses useScope(MINOR) as default.
Now the branch name, if it matches MAJOR_MINOR_X, does really increment x.