Better validation of default version strategy
This fixes #175 by running the selector() on the default strategy
before trying to use it. This should avoid messages like "Stage
final is not one of [dev] allowed for strategy development". This
could theoretically break some workflows, but in practice your
selector method should be consistent with the requirements of
the infer method. The initial message when the default can't
be selected will still be vague, but it will direct you to
rerun the build with --info which should log more information.