gh-pages only cleans up target
The pages CopySpec defaults to targeting the root of the repo,
but in cases like documentation, you may want to target a subfolder
such as one using the version in the name. This can already be done
with the into method on the CopySpec. With this change the into
setting of the pages CopySpec is also what controls which directory
is cleaned up before the copy. This allows multiple versions of doc
to exist in the repo without having the be kept in source or generated
each time. This fixes #94.