Default version strategies are undeprecated
This fixes #182.

The change for #180 was short-sighted and ignored
the reason default strategies existed in the first
place. This new fix puts the old behavior back while
still preserving the slightly better error messages
introduced in 1.5.0.

This solution is not pretty, and I'd like to use something else
but this will do for now (probably until 2.0.0).